requirements.txt
wheel
redis
sqlalchemy[asyncio]
fastapi[all]
asyncpg
python-jose[cryptography]
pytest
pydantic
passlib[bcrypt]
psycopg2-binary
pytz
pandas
requests
timezonefinder
gunicorn
pytest-cov
motor
python-jose[cryptography]
